fix: CPU metric has value wrong value

+++ src/perf/sample/writer.cpp
@@ -177,7 +177,8 @@ bool Writer::handle(const Reader::RecordSwitchType* context_switch)
                            is_switch_out);
 
     cpuid_metric_event_.timestamp(tp);
-    cpuid_metric_event_.raw_values()[0] = is_switch_out ? -1 : context_switch->cpu;
+    cpuid_metric_event_.raw_values()[0] =
+        is_switch_out ? -1 : static_cast<std::int64_t>(context_switch->cpu);
     otf2_writer_ << cpuid_metric_event_;
 
     return false;
